@@ -67,11 +67,11 @@ public class HistogramDatasetDto {
                             String.valueOf(
                                     switch (factor) {
                                         case matchHla -> sr.getTransplant().getMatchHLA();
-                                        case mismatchHla -> sr.getTransplant().getMismatchHLA();
+                                        case mismatchHla -> sr.getTransplant().isMismatchHLA();
                                         case antigen -> sr.getTransplant().getAntigen();
                                         case allele -> sr.getTransplant().getAllele();
                                         case groupHla -> sr.getTransplant().getGroup1HLA();
-                                        case postRelapse -> sr.getTransplant().getPostRelapse();
+                                        case postRelapse -> sr.getTransplant().isPostRelapse();
                                         case risk_group -> sr.getTransplant().getRecipient().getRiskGroup();
                                         case cd34perKg -> sr.getTransplant().getCD34perKg();
                                         case cd3perKg -> sr.getTransplant().getCD3perKg();
