@@ -1,9 +1,16 @@
 package com.edu.transplantdataapi.entity.transplantdata;
 
+import lombok.*;
+
 import javax.persistence.*;
 
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
 @Entity
-public class Donor{
+public class Donor {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
@@ -14,40 +21,4 @@ public class Donor{
     private Patient patient;
 
     private String stemCellSource;
-
-
-    public Donor(Patient patient, String stemCellSource) {
-        this.patient = patient;
-        this.stemCellSource = stemCellSource;
-    }
-
-    public Donor() {
-
-    }
-
-    public Patient getPatient() {
-        return patient;
-    }
-
-    public void setPatient(Patient patient) {
-        this.patient = patient;
-    }
-
-    public String getStemCellSource() {
-        return stemCellSource;
-    }
-
-    public void setStemCellSource(String stemCellSource) {
-        this.stemCellSource = stemCellSource;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-
-    }
+}
