@@ -1,7 +1,14 @@
 package com.edu.transplantdataapi.entity.transplantdata;
 
+import lombok.*;
+
 import javax.persistence.*;
 
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
 @Entity
 public class Recipient{
 
@@ -17,78 +24,4 @@ public class Recipient{
     private String disease;
     private String diseaseGroup;
     private String riskGroup;
-
-    public Recipient(Patient patient,
-                     String bloodRh,
-                     double bodyMass,
-                     String disease,
-                     String diseaseGroup,
-                     String riskGroup) {
-        this.patient = patient;
-        this.bloodRh = bloodRh;
-        this.bodyMass = bodyMass;
-        this.disease = disease;
-        this.diseaseGroup = diseaseGroup;
-        this.riskGroup = riskGroup;
-    }
-
-    public Recipient() {
-
-    }
-
-    public double getBodyMass() {
-        return bodyMass;
-    }
-
-    public void setBodyMass(double bodyMass) {
-        this.bodyMass = bodyMass;
-    }
-
-    public String getDisease() {
-        return disease;
-    }
-
-    public void setDisease(String disease) {
-        this.disease = disease;
-    }
-
-    public String getDiseaseGroup() {
-        return diseaseGroup;
-    }
-
-    public void setDiseaseGroup(String diseaseGroup) {
-        this.diseaseGroup = diseaseGroup;
-    }
-
-    public String getRiskGroup() {
-        return riskGroup;
-    }
-
-    public void setRiskGroup(String riskGroup) {
-        this.riskGroup = riskGroup;
-    }
-
-    public String getBloodRh() {
-        return bloodRh;
-    }
-
-    public void setBloodRh(String bloodRh) {
-        this.bloodRh = bloodRh;
-    }
-
-    public Patient getPatient() {
-        return patient;
-    }
-
-    public void setPatient(Patient patient) {
-        this.patient = patient;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
 }
